More memory leak fixes
Really, though...just call Log_Info(foo.c_str()); and be done with it...
