+++ Source/SudoMod.vcxproj
@@ -108,6 +108,7 @@
     <ClInclude Include="keyboard.h" />
     <ClInclude Include="Logger.h" />
     <ClInclude Include="Memory.h" />
+    <ClInclude Include="StringToChar.h" />
     <ClInclude Include="UserInterface.h" />
     <ClInclude Include="nativeCaller.h" />
     <ClInclude Include="NativeInvoker.h" />
